refactor  a bit
